moves tests around and removes tpep folder
