+++ frontendDriverInterfaceSupported.json
@@ -1,6 +1,7 @@
 {
         "_comment": "contains a list of frontend-driver interfaces branch names that this core supports",
         "versions": [
-                "1.17"
+                "1.17",
+                "2.0"
         ]
 }
