+++ recipe/dashboard/testingUtils.go
@@ -20,7 +20,6 @@ import (
 
 	"github.com/supertokens/supertokens-golang/recipe/multitenancy"
 	"github.com/supertokens/supertokens-golang/recipe/session"
-	"github.com/supertokens/supertokens-golang/recipe/thirdpartyemailpassword"
 
 	"github.com/supertokens/supertokens-golang/supertokens"
 	"github.com/supertokens/supertokens-golang/test/unittesting"
@@ -30,7 +29,6 @@ func resetAll() {
 	supertokens.ResetForTest()
 	ResetForTest()
 	session.ResetForTest()
-	thirdpartyemailpassword.ResetForTest()
 	multitenancy.ResetForTest()
 }
