+++ recipe/dashboard/userPasswordPut_test.go
@@ -8,14 +8,12 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/supertokens/supertokens-golang/recipe/dashboard/dashboardmodels"
-	"github.com/supertokens/supertokens-golang/recipe/thirdpartyemailpassword"
+	"github.com/supertokens/supertokens-golang/recipe/emailpassword"
 	"github.com/supertokens/supertokens-golang/supertokens"
 	"github.com/supertokens/supertokens-golang/test/unittesting"
 )
 
 /*
-- Initialise with thirdpartyemailpassword and provide no custom form fields
-- Create an emailpassword user using the thirdpartyemailpassword recipe
 - Try to change the password of the user
 - Should result in no errors
 - Sign in with new password
@@ -35,7 +33,7 @@ func TestThatUpdatingPasswordWithNoSignUpFeatureInTPEPWorks(t *testing.T) {
 			WebsiteDomain: "supertokens.io",
 		},
 		RecipeList: []supertokens.Recipe{
-			thirdpartyemailpassword.Init(nil),
+			emailpassword.Init(nil),
 			Init(&dashboardmodels.TypeInput{
 				ApiKey: "testapikey",
 			}),
@@ -54,7 +52,7 @@ func TestThatUpdatingPasswordWithNoSignUpFeatureInTPEPWorks(t *testing.T) {
 	testServer := httptest.NewServer(supertokens.Middleware(mux))
 	defer testServer.Close()
 
-	signupResponse, err := thirdpartyemailpassword.EmailPasswordSignUp("public", "testing@supertokens.com", "abcd1234")
+	signupResponse, err := emailpassword.SignUp("public", "testing@supertokens.com", "abcd1234")
 	if err != nil {
 		t.Error(err.Error())
 	}
@@ -78,7 +76,7 @@ func TestThatUpdatingPasswordWithNoSignUpFeatureInTPEPWorks(t *testing.T) {
 
 	assert.Equal(t, http.StatusOK, res.StatusCode)
 
-	signInResponse, err := thirdpartyemailpassword.EmailPasswordSignIn("public", "testing@supertokens.com", "newabcd1234")
+	signInResponse, err := emailpassword.SignIn("public", "testing@supertokens.com", "newabcd1234")
 	if err != nil {
 		t.Error(err.Error())
 	}
