+++ recipe/emailpassword/signoutFeature_test.go
@@ -13,7 +13,7 @@
  * under the License.
  */
 
-package thirdpartyemailpassword
+package emailpassword
 
 import (
 	"bytes"
@@ -26,8 +26,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/supertokens/supertokens-golang/recipe/session"
 	"github.com/supertokens/supertokens-golang/recipe/session/sessmodels"
-	"github.com/supertokens/supertokens-golang/recipe/thirdparty/tpmodels"
-	"github.com/supertokens/supertokens-golang/recipe/thirdpartyemailpassword/tpepmodels"
 	"github.com/supertokens/supertokens-golang/supertokens"
 	"github.com/supertokens/supertokens-golang/test/unittesting"
 	"gopkg.in/h2non/gock.v1"
@@ -45,11 +43,7 @@ func TestDefaultRouteShouldRevokeSession(t *testing.T) {
 			WebsiteDomain: "supertokens.io",
 		},
 		RecipeList: []supertokens.Recipe{
-			Init(&tpepmodels.TypeInput{
-				Providers: []tpmodels.ProviderInput{
-					customProvider2,
-				},
-			}),
+			Init(nil),
 			session.Init(&sessmodels.TypeInput{
 				AntiCsrf: &customAntiCsrfVal,
 				GetTokenTransferMethod: func(req *http.Request, forCreateNewSession bool, userContext supertokens.UserContext) sessmodels.TokenTransferMethod {
