--- recipe/thirdpartyemailpassword/tpep_userIdMapping_test.go
@@ -1,166 +1,0 @@
-package thirdpartyemailpassword
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/supertokens/supertokens-golang/recipe/thirdparty/tpmodels"
-	"github.com/supertokens/supertokens-golang/recipe/thirdpartyemailpassword/tpepmodels"
-	"github.com/supertokens/supertokens-golang/supertokens"
-	"github.com/supertokens/supertokens-golang/test/unittesting"
-)
-
-func initForUserIdMappingTest(t *testing.T) {
-
-	config := supertokens.TypeInput{
-		Supertokens: &supertokens.ConnectionInfo{
-			ConnectionURI: "http://localhost:8080",
-		},
-		AppInfo: supertokens.AppInfo{
-			APIDomain:     "api.supertokens.io",
-			AppName:       "SuperTokens",
-			WebsiteDomain: "supertokens.io",
-		},
-		RecipeList: []supertokens.Recipe{Init(&tpepmodels.TypeInput{
-			Providers: []tpmodels.ProviderInput{
-				{
-					Config: tpmodels.ProviderConfig{
-						ThirdPartyId: "google",
-						Clients: []tpmodels.ProviderClientConfig{
-							{
-								ClientID:     "test",
-								ClientSecret: "test-secret",
-							},
-						},
-					},
-				},
-			},
-		})},
-	}
-
-	err := supertokens.Init(config)
-	assert.NoError(t, err)
-}
-
-func TestCreateUserIdMappingUsingEmail(t *testing.T) {
-	BeforeEach()
-	unittesting.StartUpST("localhost", "8080")
-	defer AfterEach()
-
-	initForUserIdMappingTest(t)
-
-	querier, err := supertokens.GetNewQuerierInstanceOrThrowError("")
-	assert.NoError(t, err)
-
-	cdiVersion, err := querier.GetQuerierAPIVersion()
-	assert.NoError(t, err)
-
-	if unittesting.MaxVersion(cdiVersion, "2.14") == "2.14" {
-		return
-	}
-
-	signUpResponse, err := ThirdPartyManuallyCreateOrUpdateUser("public", "google", "googleID", "test@example.com")
-	assert.NoError(t, err)
-
-	externalUserId := "externalId"
-	externalUserIdInfo := "externalIdInfo"
-	createResp, err := supertokens.CreateUserIdMapping(signUpResponse.OK.User.ID, externalUserId, &externalUserIdInfo, nil)
-	assert.NoError(t, err)
-	assert.NotNil(t, createResp.OK)
-
-	{ // Using supertokens ID
-		userResp, err := GetUserById(signUpResponse.OK.User.ID)
-		assert.NoError(t, err)
-		assert.Equal(t, externalUserId, userResp.ID)
-	}
-
-	{ // Using external ID
-		userResp, err := GetUserById(externalUserId)
-		assert.NoError(t, err)
-		assert.Equal(t, externalUserId, userResp.ID)
-	}
-
-	{ // Using thirdparty info
-		userResp, err := GetUserByThirdPartyInfo("public", "google", "googleID")
-		assert.NoError(t, err)
-		assert.Equal(t, externalUserId, userResp.ID)
-	}
-}
-
-func TestEPCreateUserIdMappingGetUserById(t *testing.T) {
-	BeforeEach()
-	unittesting.StartUpST("localhost", "8080")
-	defer AfterEach()
-
-	initForUserIdMappingTest(t)
-
-	querier, err := supertokens.GetNewQuerierInstanceOrThrowError("")
-	assert.NoError(t, err)
-
-	cdiVersion, err := querier.GetQuerierAPIVersion()
-	assert.NoError(t, err)
-
-	if unittesting.MaxVersion(cdiVersion, "2.14") == "2.14" {
-		return
-	}
-
-	signUpResponse, err := EmailPasswordSignUp("public", "test@example.com", "testpass123")
-	assert.NoError(t, err)
-
-	assert.NotNil(t, signUpResponse.OK)
-
-	externalUserId := "externalId"
-	externalUserIdInfo := "externalIdInfo"
-	createResp, err := supertokens.CreateUserIdMapping(signUpResponse.OK.User.ID, externalUserId, &externalUserIdInfo, nil)
-	assert.NoError(t, err)
-	assert.NotNil(t, createResp.OK)
-
-	{ // Using supertokens ID
-		userResp, err := GetUserById(signUpResponse.OK.User.ID)
-		assert.NoError(t, err)
-		assert.Equal(t, externalUserId, userResp.ID)
-	}
-
-	{ // Using external ID
-		userResp, err := GetUserById(externalUserId)
-		assert.NoError(t, err)
-		assert.Equal(t, externalUserId, userResp.ID)
-	}
-}
-
-func TestEPCreateUserIdMappingGetUserByEmail(t *testing.T) {
-	BeforeEach()
-	unittesting.StartUpST("localhost", "8080")
-	defer AfterEach()
-
-	initForUserIdMappingTest(t)
-
-	querier, err := supertokens.GetNewQuerierInstanceOrThrowError("")
-	assert.NoError(t, err)
-
-	cdiVersion, err := querier.GetQuerierAPIVersion()
-	assert.NoError(t, err)
-
-	if unittesting.MaxVersion(cdiVersion, "2.14") == "2.14" {
-		return
-	}
-
-	signUpResponse, err := EmailPasswordSignUp("public", "test@example.com", "testpass123")
-	assert.NoError(t, err)
-
-	assert.NotNil(t, signUpResponse.OK)
-
-	externalUserId := "externalId"
-	externalUserIdInfo := "externalIdInfo"
-	createResp, err := supertokens.CreateUserIdMapping(signUpResponse.OK.User.ID, externalUserId, &externalUserIdInfo, nil)
-	assert.NoError(t, err)
-	assert.NotNil(t, createResp.OK)
-
-	userResp, err := GetUsersByEmail("public", "test@example.com")
-	assert.NoError(t, err)
-	assert.NotNil(t, userResp)
-	assert.Equal(t, 1, len(userResp))
-	for _, user := range userResp {
-		assert.Equal(t, externalUserId, user.ID)
-	}
-}
